Reduce duplication in drupal fixers through more granular bracket newline fixers.
