+++ Symfony/CS/Fixer/CurlyBracketsNewlineFixer.php
@@ -29,9 +29,6 @@ class CurlyBracketsNewlineFixer implements FixerInterface
         $content = $this->classDeclarationFix($content);
         $content = $this->functionDeclarationFix($content);
         $content = $this->anonymousFunctionsFix($content);
-        $content = $this->controlStatementsFix($content);
-        $content = $this->controlStatementContinuationFix($content);
-        $content = $this->doWhileFix($content);
 
         return $content;
     }
@@ -78,41 +75,5 @@ class CurlyBracketsNewlineFixer implements FixerInterface
     {
         // [Structure] No new line after anonymous function call
         return preg_replace('/((^|[\s\W])function\s*\(.*\))([^\n]*?) *\n[^\S\n]*{/', self::REMOVE_NEWLINE, $content);
-    }
-
-    private function controlStatementsFix($content)
-    {
-        $statements = array(
-            '\bif\s*\(.*\)',
-            '\belse\s*if\s*\(.*\)',
-            '\b(?<!\$)else\b',
-            '\bfor\s*\(.*\)',
-            '\b(?<!\$)do\b',
-            '\bwhile\s*\(.*\)',
-            '\bforeach\s*\(.*\)',
-            '\bswitch\s*\(.*\)',
-            '\b(?<!\$)try\b',
-            '\bcatch\s*\(.*\)',
-        );
-
-        // [Structure] No new line after control statements
-        return preg_replace('/((^|[\s\W])('.implode('|', $statements).'))([^\n]*?) *\n[^\S\n]*{/', self::REMOVE_NEWLINE, $content);
-    }
-
-    private function controlStatementContinuationFix($content)
-    {
-        $statements = array(
-            'catch',
-            'else',
-        );
-
-        // [Structure] No new line after control statements
-        return preg_replace('/}\s*\n\s*('.implode('|', $statements).')/', '} \\1', $content);
-    }
-
-    private function doWhileFix($content)
-    {
-        // [Structure] do...while loops are formatted like if {\n... \n} else {\n
-        return preg_replace('/(do {[\s\S]*)}\s*\n\s*while/', '\\1} while', $content);
     }
 }
